ref: what exactly is the idea behind lib folder
