+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/method';
-export * from './lib/service';
-export * from './lib/unsafe';
-export * from './lib/utils';
+export * from "./methods";
+export * from "./service";
+export * from "./unsafe";
+export * from "./utils";
